+++ src/models/danh-gia/dto/create-danh-gia.dto.ts
@@ -1,9 +1,11 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Schema } from 'mongoose';
+import { LopHoc } from '../../lop-hoc/lop-hoc.entity';
 import { MauDanhGia } from '../../mau-danh-gia/mau-danh-gia.entity';
 import { MonHoc } from '../../mon-hoc/mon-hoc.entity';
 import { NgayHoc } from '../../ngay-hoc/ngay-hoc.entity';
 import { NguoiDung } from '../../nguoi-dung/nguoi-dung.entity';
+import { ChiTietDG, ChiTietDGSchema } from '../chiTietDG.schema';
 
 export class CreateDanhGiaDto {
     @ApiProperty({
@@ -17,12 +19,21 @@ export class CreateDanhGiaDto {
 
     @ApiProperty({
         required: true,
-        name: 'nguoiDG',
+        name: 'lopDG',
         type: String,
         example: '60bxxxxxxxxxxxxx',
         description: 'Học sinh hay phụ huynh thực hiện đánh giá',
     })
-    nguoiDG: NguoiDung;
+    lopHoc: LopHoc;
+
+    @ApiProperty({
+        required: true,
+        name: 'monHoc',
+        type: String,
+        example: '60bxxxxxxxxxxxxx',
+        description: 'Mẫu đánh giá đặc thù dành cho môn học',
+    })
+    monHoc: MonHoc;
 
     @ApiProperty({
         required: true,
@@ -31,16 +42,17 @@ export class CreateDanhGiaDto {
         example: '60bxxxxxxxxxxxxx',
         description: 'Giáo viên được đánh giá',
     })
-    doiTuongDG: NguoiDung;
+    giaoVien: NguoiDung;
 
     @ApiProperty({
         required: true,
         name: 'ngayDG',
-        type: String,
-        example: '60bxxxxxxxxxxxxx',
+        default: Date.now(),
+        type: Date,
+        example: '21-12-2012',
         description: 'Ngày thực hiện đánh giá',
     })
-    ngayDG: NgayHoc;
+    ngayDG: Date;
 
     @ApiProperty({
         required: true,
@@ -58,4 +70,12 @@ export class CreateDanhGiaDto {
         description: 'Mẫu đánh giá',
     })
     mauDG: MauDanhGia;
+
+    @ApiProperty({
+        required: true,
+        name: 'chiTiet',
+        type: ChiTietDGSchema,
+        description: 'Chi tiết của đánh giá',
+    })
+    chiTiet: ChiTietDG[];
 }