+++ src/models/lop-hoc/lop-hoc.controller.ts
@@ -11,7 +11,13 @@ import {
 import { LopHocService } from './lop-hoc.service';
 import { CreateLopHocDto } from './dto/create-lop-hoc.dto';
 import { UpdateLopHocDto } from './dto/update-lop-hoc.dto';
-import { ApiBody, ApiCreatedResponse, ApiOkResponse, ApiParam, ApiTags } from '@nestjs/swagger';
+import {
+    ApiBody,
+    ApiCreatedResponse,
+    ApiOkResponse,
+    ApiParam,
+    ApiTags,
+} from '@nestjs/swagger';
 import { AuthGuard } from '../../auth.guard';
 
 @Controller('lop-hoc')
@@ -21,14 +27,18 @@ export class LopHocController {
     constructor(private readonly service: LopHocService) {}
 
     @Post()
-	@ApiBody({type: CreateLopHocDto})
+    @ApiBody({ type: CreateLopHocDto })
     @ApiCreatedResponse({ description: 'Tạo thành công' })
     async create(@Body() dto: CreateLopHocDto) {
         return await this.service.create(dto);
     }
 
     @Get()
-    @ApiOkResponse({ description: 'Trả về tất cả', isArray: true, type: 'array' })
+    @ApiOkResponse({
+        description: 'Trả về tất cả',
+        isArray: true,
+        type: 'array',
+    })
     async findAll() {
         return await this.service.findAll();
     }
@@ -40,15 +50,15 @@ export class LopHocController {
     }
 
     @Patch(':id')
-	@ApiParam({name: 'id', type: String, description: 'Mã lớp học'})
-	@ApiBody({type: UpdateLopHocDto})
+    @ApiParam({ name: 'id', type: String, description: 'Mã lớp học' })
+    @ApiBody({ type: UpdateLopHocDto })
     @ApiOkResponse({ description: 'Cập nhật thành công' })
     async update(@Param('id') id: string, @Body() dto: UpdateLopHocDto) {
         return await this.service.update(id, dto);
     }
 
     @Delete(':id')
-	@ApiParam({name: 'id', type: String, description: 'Mã lớp học'})
+    @ApiParam({ name: 'id', type: String, description: 'Mã lớp học' })
     @ApiOkResponse({ description: 'Xóa thành công' })
     async remove(@Param('id') id: string) {
         return await this.service.remove(id);