+++ src/users/hoc-sinh/hoc-sinh.service.ts
@@ -4,13 +4,15 @@ import { HSDGDto } from '../../models/danh-gia/dto/HSDG.dto';
 
 @Injectable()
 export class HocSinhService {
-    constructor(private dgSer: DanhGiaService) {}
+	constructor(private dgSer: DanhGiaService) { }
 
-    async getOne_danhGia() {
-        return await this.dgSer.findOne('60b745814bddbdc7d2df9b75');
-    }
+	async getOne_danhGia() {
+		return await this.dgSer.findOne('60b745814bddbdc7d2df9b75');
+	}
 
-    async makeReview(id: string, dto: HSDGDto) {
-        return await this.dgSer.update(id, dto);
-    }
+	async makeReview(id: string, dto: HSDGDto) {
+		const ct = { chiTiet: [] }
+		ct.chiTiet.push(dto)
+		return await this.dgSer.update(id, ct)
+	}
 }